server/pyproject.toml
[tool.poetry]
name = "capstone-project-label"
version = "1.00"
description = "Backend for label"
authors = ["AnuragAmbekar741 <anuragambekar1997@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.118.0"
strawberry-graphql = {extras = ["fastapi"], version = "^0.283.2"}
tortoise-orm = "^0.25.1"
celery = "^5.5.3"
uvicorn = "^0.37.0"
asyncpg = "^0.30.0"
redis = "^6.4.0"
python-multipart = "^0.0.20"
aerich = "^0.9.1"
pydantic = "^2.12.0"
python-jose = {extras = ["cryptography"], version = "^3.5.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
email-validator = "^2.3.0"
pydantic-settings = "^2.11.0"
python-dotenv = "^1.1.1"
google-auth = "^2.41.1"
pyjwt = "^2.10.1"
requests = "^2.32.5"
httpx = "^0.28.1"
google-auth-oauthlib = "^1.2.2"
imapclient = "^3.0.1"
langchain = "^1.0.7"
langchain-google-genai = "^3.0.3"
langchain-openai = "^1.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.2"
black = "^25.9.0"
flake8 = "^7.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.aerich]
tortoise_orm = "app.config.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."
